Clean up signing module
