+++ rustica/Cargo.toml
@@ -28,10 +28,10 @@ ring = "0.16.20"
 serde = {version = "1.0", features = ["derive"]}
 sha2 = "0.9.2"
 # For Production
-# sshcerts = { version = "0.7", features = ["x509-support"]}
+sshcerts = { version = "0.8", features = ["x509-support"]}
 # For Development
 # sshcerts = {git = "https://github.com/obelisk/sshcerts", features = ["yubikey"]}
-sshcerts = {path = "../../sshcerts", features = ["x509-support"]}
+# sshcerts = {path = "../../sshcerts", features = ["x509-support"]}
 tokio = { version = "1.0.0", features = ["full"] }
 toml = "0.5"
 tonic = { version = "0.4", features = ["tls"] }
