+++ rustica/src/signing/vault.rs
@@ -1,0 +1,4 @@
+use serde::Deserialize;
+
+#[derive(Deserialize)]
+pub struct VaultSigner {}
